// frontend/src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        CsrfBootstrapResponse: {
            /** Absolute Expires At Ms */
            absolute_expires_at_ms: number;
            /** Actor Id */
            actor_id: string;
            actor_type: components["schemas"]["SessionActorType"];
            /** Csrf Token */
            csrf_token: string;
            /** Display Name */
            display_name: string;
            /** Participant Id */
            participant_id?: string | null;
            /** Room Id */
            room_id?: string | null;
        };
        /** CurrentSessionResponse */
        CurrentSessionResponse: {
            /** Absolute Expires At Ms */
            absolute_expires_at_ms: number;
            /** Actor Id */
            actor_id: string;
            actor_type: components["schemas"]["SessionActorType"];
            /** Display Name */
            display_name: string;
            /** Participant Id */
            participant_id?: string | null;
            /** Room Id */
            room_id?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}